+++ qupath-core/src/main/java/qupath/lib/images/servers/TransformedServerBuilder.java
@@ -58,9 +58,10 @@ public class TransformedServerBuilder {
 	}
 	
 	/**
-	 * Crop an specified region based on a bounding box.
-	 * @param region
-	 * @return
+	 * Crop a specified region based on a bounding box.
+	 *
+	 * @param region the region to crop
+	 * @return this builder
 	 */
 	public TransformedServerBuilder crop(ImageRegion region) {
 		server = new CroppedImageServer(server, region);
@@ -70,8 +71,9 @@ public class TransformedServerBuilder {
 	/**
 	 * Apply an {@link AffineTransform} to the server. 
 	 * Note that the transform must be invertible, otherwise and {@link IllegalArgumentException} will be thrown.
-	 * @param transform
-	 * @return
+	 *
+	 * @param transform the transform to apply to the image
+	 * @return this builder
 	 */
 	public TransformedServerBuilder transform(AffineTransform transform) {
 		try {
@@ -83,10 +85,11 @@ public class TransformedServerBuilder {
 	}
 	
 	/**
-	 * Apply color deconvolution to the brightfield image, so that deconvolved stains behave as separate channels/
+	 * Apply color deconvolution to the brightfield image, so that deconvolved stains behave as separate channels.
+	 *
 	 * @param stains the stains to apply for color deconvolution
 	 * @param stainNumbers the indices of the stains that should be use (an array compressing values that are 1, 2 or 3); if not specified, all 3 stains will be used.
-	 * @return
+	 * @return this builder
 	 */
 	public TransformedServerBuilder deconvolveStains(ColorDeconvolutionStains stains, int...stainNumbers) {
 		server = new ColorDeconvolutionImageServer(server, stains, stainNumbers);
@@ -96,8 +99,9 @@ public class TransformedServerBuilder {
 	/**
 	 * Rearrange the channel order of an RGB image.
 	 * This is intended for cases where an image has wrongly been interpreted as RGB or BGR.
-	 * @param order
-	 * @return
+	 *
+	 * @param order a text containing the letters R, G, and B in any order
+	 * @return this builder
 	 */
 	public TransformedServerBuilder reorderRGB(String order) {
 		server = new RearrangeRGBImageServer(server, order);
@@ -106,8 +110,9 @@ public class TransformedServerBuilder {
 	
 	/**
 	 * Rotate the image, using an increment of 90 degrees.
-	 * @param rotation
-	 * @return
+	 *
+	 * @param rotation the rotation to apply
+	 * @return this image
 	 */
 	public TransformedServerBuilder rotate(Rotation rotation) {
 		server = new RotatedImageServer(server, rotation);
@@ -116,8 +121,9 @@ public class TransformedServerBuilder {
 	
 	/**
 	 * Extract specified channels for an image.
+	 *
 	 * @param channels indices (0-based) of channels to extract.
-	 * @return
+	 * @return this builder
 	 */
 	public TransformedServerBuilder extractChannels(int... channels) {
 		var transforms = new ArrayList<ColorTransform>();
@@ -130,8 +136,9 @@ public class TransformedServerBuilder {
 	
 	/**
 	 * Extract specified channels for an image.
+	 *
 	 * @param names names of channels to extract.
-	 * @return
+	 * @return this builder
 	 */
 	public TransformedServerBuilder extractChannels(String... names) {
 		var transforms = new ArrayList<ColorTransform>();
@@ -144,48 +151,42 @@ public class TransformedServerBuilder {
 	
 	/**
 	 * Perform a maximum projection of the channels.
-	 * @return
+	 *
+	 * @return this builder
 	 */
 	public TransformedServerBuilder maxChannelProject() {
-		server = new ChannelTransformFeatureServer(server,
-				List.of(ColorTransforms.createMaximumChannelTransform()));
+		server = new ChannelTransformFeatureServer(server, List.of(ColorTransforms.createMaximumChannelTransform()));
 		return this;
 	}
 	
 	/**
 	 * Perform an average (mean) projection of the channels.
-	 * @return
+	 *
+	 * @return this builder
 	 */
 	public TransformedServerBuilder averageChannelProject() {
-		server = new ChannelTransformFeatureServer(server,
-				List.of(ColorTransforms.createMeanChannelTransform()));
+		server = new ChannelTransformFeatureServer(server, List.of(ColorTransforms.createMeanChannelTransform()));
 		return this;
 	}
 	
 	/**
 	 * Perform a minimum projection of the channels.
-	 * @return
+	 *
+	 * @return this builder
 	 */
 	public TransformedServerBuilder minChannelProject() {
-		server = new ChannelTransformFeatureServer(server,
-				List.of(ColorTransforms.createMinimumChannelTransform()));
+		server = new ChannelTransformFeatureServer(server, List.of(ColorTransforms.createMinimumChannelTransform()));
 		return this;
 	}
 	
 	/**
 	 * Concatenate a collection of additional servers along the 'channels' dimension (iteration order is used).
-	 * @param additionalChannels additional servers that will be applied as channels; note that these should be 
-	 * 								of an appropriate type and dimension for concatenation.
-	 * @return
+	 *
+	 * @param additionalChannels additional servers that will be applied as channels; note that these should be
+	 *                           of an appropriate type and dimension for concatenation.
+	 * @return this builder
 	 */
 	public TransformedServerBuilder concatChannels(Collection<ImageServer<BufferedImage>> additionalChannels) {
-//		// Try to avoid wrapping more than necessary
-//		if (server instanceof ConcatChannelsImageServer) {
-//			var temp = new ArrayList<>(((ConcatChannelsImageServer)server).getAllServers());
-//			temp.addAll(additionalChannels);
-//			server = new ConcatChannelsImageServer(((ConcatChannelsImageServer)server).getWrappedServer(), temp);
-//		} else
-		
 		List<ImageServer<BufferedImage>> allChannels = new ArrayList<>(additionalChannels);
 		// Make sure that the current server is included
 		if (!allChannels.contains(server))
@@ -196,9 +197,10 @@ public class TransformedServerBuilder {
 	
 	/**
 	 * Concatenate additional servers along the 'channels' dimension.
-	 * @param additionalChannels additional servers from which channels will be added; note that the servers should be 
-	 * 								of an appropriate type and dimension for concatenation.
-	 * @return
+	 *
+	 * @param additionalChannels additional servers from which channels will be added; note that the servers should be
+	 *                           of an appropriate type and dimension for concatenation.
+	 * @return this builder
 	 */
 	public TransformedServerBuilder concatChannels(ImageServer<BufferedImage>... additionalChannels) {
 		for (var temp : additionalChannels) {
@@ -268,10 +270,31 @@ public class TransformedServerBuilder {
 		this.server = new NormalizedImageServer(server, normalizer);
 		return this;
 	}
+
+	/**
+	 * Apply color transforms to the image.
+	 *
+	 * @param transforms the transforms to apply
+	 * @return this builder
+	 */
+	public TransformedServerBuilder applyColorTransforms(Collection<? extends ColorTransform> transforms) {
+		server = new ChannelTransformFeatureServer(server, new ArrayList<>(transforms));
+		return this;
+	}
+
+	/**
+	 * Apply color transforms to the image.
+	 *
+	 * @param transforms the transforms to apply
+	 * @return this builder
+	 */
+	public TransformedServerBuilder applyColorTransforms(ColorTransform... transforms) {
+		return applyColorTransforms(Arrays.asList(transforms));
+	}
+
 	
 	/**
 	 * Get the {@link ImageServer} that applies all the requested transforms.
-	 * @return
 	 */
 	public ImageServer<BufferedImage> build() {
 		return server;