+++ qupath-core/src/main/java/qupath/lib/images/writers/TileExporter.java
@@ -537,7 +537,7 @@ public class TileExporter  {
 						continue;
 					}
 				} else if (imageData != null) {
-					if (!imageData.getHierarchy().getObjectsForRegion(PathAnnotationObject.class, r, null)
+					if (!imageData.getHierarchy().getAnnotationsForRegion(r, null)
 							.stream().anyMatch(p -> RoiTools.intersectsRegion(p.getROI(), r))) {
 						iterator.remove();
 						continue;