+++ qupath-core/src/main/java/qupath/lib/projects/DefaultProject.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2020 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2024 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -63,8 +63,8 @@ import qupath.lib.common.GeneralTools;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.ImageData.ImageType;
 import qupath.lib.images.servers.ImageServer;
-import qupath.lib.images.servers.ServerTools;
 import qupath.lib.images.servers.ImageServerBuilder.ServerBuilder;
+import qupath.lib.images.servers.ImageServerMetadata;
 import qupath.lib.io.GsonTools;
 import qupath.lib.io.PathIO;
 import qupath.lib.objects.PathObject;
@@ -475,7 +475,9 @@ class DefaultProject implements Project<BufferedImage> {
 				setDescription(description);
 			
 			if (metadataMap != null)
-				metadata.putAll(metadataMap);		
+				metadata.putAll(metadataMap);
+
+			writeServerBuilder();
 		}
 		
 		DefaultProjectImageEntry(final DefaultProjectImageEntry entry) {
@@ -526,6 +528,8 @@ class DefaultProject implements Project<BufferedImage> {
 				Files.copy(entry.getImageDataPath(), getImageDataPath(), StandardCopyOption.REPLACE_EXISTING);
 			if (Files.exists(entry.getDataSummaryPath()))
 				Files.copy(entry.getDataSummaryPath(), getDataSummaryPath(), StandardCopyOption.REPLACE_EXISTING);
+			if (Files.exists(entry.getServerPath()))
+				Files.copy(entry.getServerPath(), getServerPath(), StandardCopyOption.REPLACE_EXISTING);
 			if (getThumbnail() == null && Files.exists(entry.getThumbnailPath()))
 				Files.copy(entry.getThumbnailPath(), getThumbnailPath(), StandardCopyOption.REPLACE_EXISTING);
 		}
@@ -561,6 +565,8 @@ class DefaultProject implements Project<BufferedImage> {
 			var builderBefore = serverBuilder;
 			serverBuilder = serverBuilder.updateURIs(replacements);
 			boolean changes = builderBefore != serverBuilder;
+			if (changes)
+				writeServerBuilder();
 			return changes;
 		}
 		
@@ -688,20 +694,11 @@ class DefaultProject implements Project<BufferedImage> {
 		@Override
 		public synchronized ImageData<BufferedImage> readImageData() throws IOException {
 			Path path = getImageDataPath();
-			ImageServer<BufferedImage> server;
-			try {
-				server = getServerBuilder().build();
-			} catch (IOException e) {
-				throw e;
-			} catch (Exception e) {
-				throw new IOException(e);
-			}
-			if (server == null)
-				return null;
 			ImageData<BufferedImage> imageData = null;
+			// TODO: Consider whether we can set the image name for the lazy-loaded server
 			if (Files.exists(path)) {
 				try (var stream = Files.newInputStream(path)) {
-					imageData = PathIO.readImageData(stream, null, server, BufferedImage.class);
+					imageData = PathIO.readLazyImageData(stream, getServerBuilder(), BufferedImage.class);
 					imageData.setLastSavedPath(path.toString(), true);
 				} catch (Exception e) {
 					logger.error("Error reading image data from " + path, e);
@@ -713,7 +710,7 @@ class DefaultProject implements Project<BufferedImage> {
 				var pathBackup = getBackupImageDataPath();
 				if (Files.exists(pathBackup)) {
 					try (var stream = Files.newInputStream(pathBackup)) {
-						imageData = PathIO.readImageData(stream, null, server, BufferedImage.class);
+						imageData = PathIO.readLazyImageData(stream, getServerBuilder(), BufferedImage.class);
 						imageData.setLastSavedPath(pathBackup.toString(), true);
 						logger.warn("Restored previous ImageData from {}", pathBackup);
 					} catch (IOException e) {
@@ -723,13 +720,16 @@ class DefaultProject implements Project<BufferedImage> {
 			}
 			
 			if (imageData == null)
-				imageData = new ImageData<>(server);
-			// Ensure the names match
-			var name = getOriginalImageName();
-			if (name != null)
-				ServerTools.setImageName(server, name);
+				imageData = new ImageData<>(getServerBuilder(), new PathObjectHierarchy(), ImageType.UNSET);
 			imageData.setProperty(IMAGE_ID, getFullProjectEntryID()); // Required to be able to test for the ID later
 			imageData.setChanged(false);
+			// I don't like it either - but we want to ensure that the server name matches with the image entry name.
+			// This can trigger lazy-loading of the server, but it's necessary to ensure that the server name is correct.
+			imageData.updateServerMetadata(
+					new ImageServerMetadata.Builder(imageData.getServerMetadata())
+							.name(getImageName())
+							.build()
+			);
 			return imageData;
 		}
 
@@ -777,17 +777,8 @@ class DefaultProject implements Project<BufferedImage> {
 			var currentServerBuilder = server.getBuilder();
 			if (currentServerBuilder != null && !currentServerBuilder.equals(this.serverBuilder)) {
 				this.serverBuilder = currentServerBuilder;
-				// Write the server - it isn't used, but it may enable us to rebuild the server from the data directory
-				// if the project is lost.
-				// Note that before v0.5.0, this actually wrote the server builder - but this was missing type
-				// information, so recovery of the actual server was difficult.
-				var pathServer = getServerPath();
-				try (var out = Files.newBufferedWriter(pathServer, StandardCharsets.UTF_8)) {
-					GsonTools.getInstance().toJson(server, out);
-				} catch (Exception e) {
-					logger.warn("Unable to write server to {}", pathServer);
-					Files.deleteIfExists(pathServer);
-				}
+				writeServerBuilder();
+				// This ensures that the metadata is updated in the project file
 //				syncChanges();
 			}
 			
@@ -796,6 +787,22 @@ class DefaultProject implements Project<BufferedImage> {
 				GsonTools.getInstance().toJson(new ImageDataSummary(imageData, timestamp), out);
 			}			
 
+		}
+
+		private void writeServerBuilder() throws IOException {
+			// Write the server - it isn't used, but it may enable us to rebuild the server from the data directory
+			// if the project is lost.
+			// Note that before v0.5.0, this actually wrote the server builder - but this was missing type
+			// information, so recovery of the actual server was difficult.
+			getEntryPath(true); // Ensure the directory exists
+			var pathServer = getServerPath();
+			try (var out = Files.newBufferedWriter(pathServer, StandardCharsets.UTF_8)) {
+				// Important to specify the class as ServerBuilder, so that the type adapter writes the type!
+				GsonTools.getInstance(true).toJson(this.serverBuilder, ServerBuilder.class, out);
+			} catch (Exception e) {
+				logger.warn("Unable to write server to {}", pathServer);
+				Files.deleteIfExists(pathServer);
+			}
 		}
 
 		@Override
@@ -864,8 +871,18 @@ class DefaultProject implements Project<BufferedImage> {
 			resetCachedThumbnail();
 			getEntryPath(true);
 			var path = getThumbnailPath();
-			try (var stream = Files.newOutputStream(path)) {
-				ImageIO.write(img, "JPEG", stream);
+			if (img == null) {
+				// Reset the thumbnail
+				if (Files.exists(path)) {
+					logger.debug("Deleting thumbnail for {}", path);
+					Files.delete(path);
+				}
+			} else {
+				// Save the thumbnail
+				try (var stream = Files.newOutputStream(path)) {
+					logger.debug("Writing thumbnail to {}", path);
+					ImageIO.write(img, "JPEG", stream);
+				}
 			}
 		}